% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifExtraTrees.R
\docType{data}
\name{LearnerClassifExtraTrees}
\alias{LearnerClassifExtraTrees}
\alias{mlr_learners_classif.extratrees}
\title{Classification ExtraTrees Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifExtraTrees
}
\description{
A \link[mlr3:LearnerClassif]{mlr3::LearnerClassif} for a classification extratree implemented in extraTree::extraTree()] in package \CRANpkg{extraTree}.
}
\keyword{datasets}
