% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrExtraTrees.R
\docType{data}
\name{LearnerRegrExtraTrees}
\alias{LearnerRegrExtraTrees}
\alias{mlr_learners_regr.extratrees}
\title{Regression ExtraTrees Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrExtraTrees
}
\description{
A \link[mlr3:LearnerRegr]{mlr3::LearnerRegr} for a regression extratree implemented in extraTree::extraTree()] in package \CRANpkg{extraTree}.
}
\keyword{datasets}
