% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.extratrees-package}
\alias{mlr3learners.extratrees}
\alias{mlr3learners.extratrees-package}
\title{mlr3learners.extratrees: extraTrees learner from extraTrees package for mlr3}
\description{
extraTrees learner from extraTrees package for
    mlr3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr3learners/mlr3learners.extratrees}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de}

}
